requirements.txt
anyio==3.3.4
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
idna==3.3
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3
urllib3==1.26.7
